+++ src/controllers/console/Oauth2ClientController.php
@@ -4,11 +4,16 @@ namespace rhertogh\Yii2Oauth2Server\controllers\console;
 
 use rhertogh\Yii2Oauth2Server\controllers\console\base\Oauth2BaseConsoleController;
 use rhertogh\Yii2Oauth2Server\controllers\console\client\Oauth2CreateClientAction;
+use rhertogh\Yii2Oauth2Server\controllers\console\client\Oauth2DeleteClientAction;
+use rhertogh\Yii2Oauth2Server\controllers\console\client\Oauth2ListClientsAction;
 use rhertogh\Yii2Oauth2Server\controllers\console\client\Oauth2SetClientSecretAction;
+use rhertogh\Yii2Oauth2Server\controllers\console\client\Oauth2UpdateClientAction;
+use rhertogh\Yii2Oauth2Server\controllers\console\client\Oauth2ViewClientAction;
 use yii\helpers\ArrayHelper;
 
 class Oauth2ClientController extends Oauth2BaseConsoleController
 {
+    public $defaultAction = 'list';
     /**
      * @var string|null
      */
@@ -42,7 +47,19 @@ class Oauth2ClientController extends Oauth2BaseConsoleController
     /**
      * @var string|null
      */
+    public $allowVariableRedirectUriQuery = null;
+
+    /**
+     * @var string|null
+     */
     public $secret = null;
+
+    /**
+     * @var string|null A PHP date/time string or, when starting with 'P', a PHP date interval string.
+     * @see https://www.php.net/manual/en/datetime.formats.php
+     * @see https://www.php.net/manual/en/dateinterval.construct.php
+     */
+    public $oldSecretValidUntil = null;
 
     /**
      * @var string|null
@@ -52,14 +69,42 @@ class Oauth2ClientController extends Oauth2BaseConsoleController
     /**
      * @var string|null
      */
+    public $scopeAccess = null;
+
+    /**
+     * @var string|null
+     */
     public $endUsersMayAuthorizeClient = null;
 
     /**
-     * @var string|null A PHP date/time string or, when starting with 'P', a PHP date interval string.
-     * @see https://www.php.net/manual/en/datetime.formats.php
-     * @see https://www.php.net/manual/en/dateinterval.construct.php
+     * @var string|null
      */
-    public $oldSecretValidUntil = null;
+    public $userAccountSelection = null;
+
+    /**
+     * @var string|null
+     */
+    public $isAuthCodeWithoutPkceAllowed = null;
+
+    /**
+     * @var string|null
+     */
+    public $skipAuthorizationIfScopeIsAllowed = null;
+
+    /**
+     * @var string|null
+     */
+    public $logoUri = null;
+
+    /**
+     * @var string|null
+     */
+    public $termsOfServiceUri = null;
+
+    /**
+     * @var string|null
+     */
+    public $contacts = null;
 
     /**
      * @inheritDoc
@@ -73,10 +118,18 @@ class Oauth2ClientController extends Oauth2BaseConsoleController
                 'name',
                 'type',
                 'redirectURIs',
+                'allowVariableRedirectUriQuery',
                 'grantTypes',
                 'secret',
                 'scopes',
+                'scopeAccess',
                 'endUsersMayAuthorizeClient',
+                'userAccountSelection',
+                'isAuthCodeWithoutPkceAllowed',
+                'skipAuthorizationIfScopeIsAllowed',
+                'logoUri',
+                'termsOfServiceUri',
+                'contacts',
             ];
         } elseif ($actionID == 'set-secret') {
             $options = [
@@ -95,8 +148,12 @@ class Oauth2ClientController extends Oauth2BaseConsoleController
     public function actions()
     {
         return [
+            'list' => Oauth2ListClientsAction::class,
+            'view' => Oauth2ViewClientAction::class,
             'create' => Oauth2CreateClientAction::class,
+            'update' => Oauth2UpdateClientAction::class,
             'set-secret' => Oauth2SetClientSecretAction::class,
+            'delete' => Oauth2DeleteClientAction::class,
         ];
     }
 }