+++ src/controllers/console/Oauth2DebugController.php
@@ -7,7 +7,11 @@ use rhertogh\Yii2Oauth2Server\controllers\console\debug\Oauth2DebugConfigAction;
 
 class Oauth2DebugController extends Oauth2BaseConsoleController
 {
+    /**
+     * @inheritDoc
+     */
     public $defaultAction = 'config';
+
     /**
      * @inheritDoc
      */