+++ src/interfaces/components/repositories/Oauth2ClientRepositoryInterface.php
@@ -5,6 +5,7 @@ namespace rhertogh\Yii2Oauth2Server\interfaces\components\repositories;
 use League\OAuth2\Server\Repositories\ClientRepositoryInterface;
 use rhertogh\Yii2Oauth2Server\interfaces\components\repositories\base\Oauth2ModelRepositoryInterface;
 use rhertogh\Yii2Oauth2Server\interfaces\models\Oauth2ClientInterface;
+use yii\base\InvalidConfigException;
 
 interface Oauth2ClientRepositoryInterface extends
     Oauth2ModelRepositoryInterface,
@@ -17,4 +18,11 @@ interface Oauth2ClientRepositoryInterface extends
      */
     public function getClientEntity($clientIdentifier);
     # endregion
+
+    /**
+     * @param array $filter
+     * @return Oauth2ClientInterface[]
+     * @throws InvalidConfigException
+     */
+    public function getAllClients($filter = []);
 }